requirements.txt
slack_sd==3.4.0
spotipy==2.4.4

